pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dropbox-automation"
version = "1.0.0"
description = "An application that processes PDF files in Dropbox using Claude AI"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "picane", email = "piclane@gmail.com" }
]

dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "dropbox>=12.0.2",
    "anthropic>=0.5.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "PyMuPDF>=1.25.5"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.hatch.build.targets.wheel]
packages = ["src/dropbox_pdf_processor"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]
